iOS GC fixes for VM (#848)
* Add explicit 64->32 bit casts to silence iOS warnings

* Fix GC memory allocation on iOS

* Check for OOM on iOS

+++ libs/base/gc.cpp
@@ -654,7 +654,7 @@ void *gcAllocate(int numbytes) {
             auto vt = p->vtable;
             if (!IS_FREE(vt))
                 oops(43);
-            int left = VAR_BLOCK_WORDS(vt) - numwords;
+            int left = (int)(VAR_BLOCK_WORDS(vt) - numwords);
             VVLOG("%p %d - %d = %d", (void*)vt, (int)VAR_BLOCK_WORDS(vt), (int)numwords, left);
             if (left >= 0) {
                 auto nf = (RefBlock *)((void **)p + numwords);
